+++ media/lua/client/UI/CHC_window.lua
@@ -106,6 +106,7 @@ function CHC_window:onResize()
 end
 
 function CHC_window:render()
+    if self.isCollapsed then return end;
     ISCollapsableWindow.render(self);
 end