+++ CHANGELOG.md
@@ -1,6 +1,29 @@
 # CraftHelper updates
 
-# 0.1 (27.03.2022)
+# 1.2 (28.03.2022)
+
+### New features
+
+- Added ability to favorite recipes from craft helper window
+  - semi-synchronized with crafting window (need to re-open crafthelper or change category)
+
+### Changes
+
+- Changed event used to load all recipes (OnGameBoot -> OnGameStart)
+- Handle config creation
+
+# 1.1 (27.03.2022)
+
+### New features
+
+- Added configuration file
+- Added saving for various properties on window close
+
+### Changes
+
+- Refactored codebase
+
+# 1.0 (27.03.2022)
 
 ### New features